LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/navigate-top-to-aboutblank.https-expected.txt

Non-initial empty documents (about:blank) should inherit their cross-origin-opener-policy from the navigation's initiator top level document, if the initiator and its top level document are same-origin, or default (to unsafe-none) otherwise.

Create the opener popup with a given COOP openerCOOP.
Add iframe to the opener popup that is either same-origin or cross-origin.
Opener opens a new window, to a network document with the same origin and COOP value as opener.
Opener's iframe navigates its parent frame (opener) to about:blank.
Verify the openee still has access to its opener.

FAIL Navigate top to about:blank from iframe with opener COOP: |header(Cross-Origin-Opener-Policy,same-origin), iframe origin: https://localhost:9443 assert_equals: expected "about:blank loaded" but got "about:blank NOT loaded"
FAIL Navigate top to about:blank from iframe with opener COOP: |header(Cross-Origin-Opener-Policy,same-origin), iframe origin: https://127.0.0.1:9443 assert_equals: expected "about:blank loaded" but got "about:blank NOT loaded"
FAIL Navigate top to about:blank from iframe with opener COOP: |header(Cross-Origin-Opener-Policy,same-origin-allow-popups), iframe origin: https://127.0.0.1:9443 assert_equals: expected "about:blank loaded" but got "about:blank NOT loaded"
